Split handlers into their individual files
Eventually, this will allow for us to cover further namespaces, such as CCS or NUS proper.
