+++ main.go
@@ -24,9 +24,6 @@ import (
 	"io/ioutil"
 	"log"
 	"net/http"
-	"strconv"
-	"strings"
-	"time"
 
 	_ "github.com/go-sql-driver/mysql"
 )
@@ -41,7 +38,7 @@ const (
 // checkError makes error handling not as ugly and inefficient.
 func checkError(err error) {
 	if err != nil {
-		log.Fatalf("WiiSOAP forgot how to drive and suddenly crashed! Reason: %s\n", err.Error())
+		log.Fatalf("WiiSOAP forgot how to drive and suddenly crashed! Reason: %v\n", err)
 	}
 }
 
@@ -69,313 +66,12 @@ func main() {
 
 	// Start the HTTP server.
 	fmt.Printf("Starting HTTP connection (%s)...\nNot using the usual port for HTTP?\nBe sure to use a proxy, otherwise the Wii can't connect!\n", CON.Address)
-	http.HandleFunc("/ecs/services/ECommerceSOAP", handler) // each request calls handler
+
+	// These following endpoints don't have to match what the official WSC have.
+	// However, semantically, it feels proper.
+	http.HandleFunc("/ecs/services/ECommerceSOAP", ecsHandler)              // For ECS operations
+	http.HandleFunc("/ias/services/IdentityAuthenticationSOAP", iasHandler) // For IAS operations
 	log.Fatal(http.ListenAndServe(CON.Address, nil))
 
-	// From here on out, all special cool things should go into the handler function.
-}
-
-func handler(w http.ResponseWriter, r *http.Request) {
-	// Figure out the action to handle via header.
-	action := r.Header.Get("SOAPAction")
-	action = parseAction(action, "ecs")
-
-	// Get a sexy new timestamp to use.
-	timestampnano := strconv.FormatInt(time.Now().UTC().Unix(), 10)
-	timestamp := timestampnano + "000"
-
-	fmt.Println("[!] Incoming request.")
-	body, err := ioutil.ReadAll(r.Body)
-	if err != nil {
-		http.Error(w, "Error reading request body...", http.StatusInternalServerError)
-	}
-
-	// The switch converts the HTTP Body of the request into a string. There is no need to convert the cases to byte format.
-	switch action {
-	// TODO: Make the case functions cleaner. (e.g. Should the response be a variable?)
-	// TODO: Update the responses so that they query the SQL Database for the proper information (e.g. Device Code, Token, etc).
-
-	case "CheckDeviceStatus":
-		fmt.Println("CDS.")
-		CDS := CDS{}
-		if err = xml.Unmarshal(body, &CDS); err != nil {
-			fmt.Println("...or not. Bad or incomplete request. (End processing.)")
-			fmt.Fprint(w, "You need to POST some SOAP from WSC if you wanna get some, honey. ;3")
-			return
-		}
-		fmt.Println(CDS)
-		fmt.Println("The request is valid! Responding...")
-		fmt.Fprintf(w, `<?xml version="1.0" encoding="utf-8"?>
-<soapenv:Envelope xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/" 
-	  	xmlns:xsd="http://www.w3.org/2001/XMLSchema" 
-	  	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-	<soapenv:Body>
-		<CheckDeviceStatusResponse xmlns="urn:ecs.wsapi.broadon.com">
-			<Version>%s</Version>
-			<DeviceId>%s</DeviceId>
-			<MessageId>%s</MessageId>
-			<TimeStamp>%s</TimeStamp>
-			<ErrorCode>0</ErrorCode>
-			<ServiceStandbyMode>false</ServiceStandbyMode>
-			<Balance>
-				<Amount>2018</Amount>
-				<Currency>POINTS</Currency>
-			</Balance>
-			<ForceSyncTime>0</ForceSyncTime>
-			<ExtTicketTime>%s</ExtTicketTime>
-			<SyncTime>%s</SyncTime>
-		</CheckDeviceStatusResponse>
-	</soapenv:Body>
-</soapenv:Envelope>`, CDS.Version, CDS.DeviceID, CDS.MessageID, timestamp, timestamp, timestamp)
-		fmt.Println("Delivered response!")
-
-	case "NotifiedETicketsSynced":
-		fmt.Println("NETS")
-		NETS := NETS{}
-		if err = xml.Unmarshal(body, &NETS); err != nil {
-			fmt.Println("...or not. Bad or incomplete request. (End processing.)")
-			fmt.Fprint(w, "This is a disgusting request, but 20 dollars is 20 dollars. ;3")
-			fmt.Printf("error: %v", err)
-			return
-		}
-		fmt.Println(NETS)
-		fmt.Println("The request is valid! Responding...")
-		fmt.Fprintf(w, `<?xml version="1.0" encoding="utf-8"?>
-<soapenv:Envelope xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/" 
-			xmlns:xsd="http://www.w3.org/2001/XMLSchema" 
-			xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-		<soapenv:Body>
-			<NotifyETicketsSyncedResponse xmlns="urn:ecs.wsapi.broadon.com">
-				<Version>%s</Version>
-				<DeviceId>%s</DeviceId>
-				<MessageId>%s</MessageId>
-				<TimeStamp>%s</TimeStamp>
-				<ErrorCode>0</ErrorCode>
-				<ServiceStandbyMode>false</ServiceStandbyMode>
-			</NotifyETicketsSyncedResponse>
-	</soapenv:Body>
-</soapenv:Envelope>`, NETS.Version, NETS.DeviceID, NETS.MessageID, timestamp)
-		fmt.Println("Delivered response!")
-
-	case "ListETickets":
-		fmt.Println("LET")
-		LET := LET{}
-		if err = xml.Unmarshal(body, &LET); err != nil {
-			fmt.Println("...or not. Bad or incomplete request. (End processing.)")
-			fmt.Fprint(w, "This is a disgusting request, but 20 dollars is 20 dollars. ;3")
-			fmt.Printf("error: %v", err)
-			return
-		}
-		fmt.Println(LET)
-		fmt.Println("The request is valid! Responding...")
-		fmt.Fprintf(w, `<?xml version="1.0" encoding="utf-8"?>
-<soapenv:Envelope xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/" 
-		xmlns:xsd="http://www.w3.org/2001/XMLSchema" 
-		xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-	<soapenv:Body>
-		<ListETicketsResponse xmlns="urn:ecs.wsapi.broadon.com">
-			<Version>%s</Version>
-			<DeviceId>%s</DeviceId>
-			<MessageId>%s</MessageId>
-			<TimeStamp>%s</TimeStamp>
-			<ErrorCode>0</ErrorCode>
-			<ServiceStandbyMode>false</ServiceStandbyMode>
-			<ForceSyncTime>0</ForceSyncTime>
-			<ExtTicketTime>%s</ExtTicketTime>
-			<SyncTime>%s</SyncTime>
-		</ListETicketsResponse>
-	</soapenv:Body>
-</soapenv:Envelope>`, LET.Version, LET.DeviceID, LET.MessageID, timestamp, timestamp, timestamp)
-		fmt.Println("Delivered response!")
-
-	case "PurchaseTitle":
-		fmt.Println("PT")
-		PT := PT{}
-		if err = xml.Unmarshal(body, &PT); err != nil {
-			fmt.Println("...or not. Bad or incomplete request. (End processing.)")
-			fmt.Fprint(w, "if you wanna fun time, its gonna cost ya extra sweetie. ;3")
-			fmt.Printf("Error: %s", err.Error())
-			return
-		}
-		fmt.Println(PT)
-		fmt.Println("The request is valid! Responding...")
-		fmt.Fprintf(w, `<?xml version="1.0" encoding="utf-8"?>
-<soapenv:Envelope	xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/" 
-		xmlns:xsd="http://www.w3.org/2001/XMLSchema" 
-		xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-	<soapenv:Body>
-		<PurchaseTitleResponse xmlns="urn:ecs.wsapi.broadon.com">
-			<Version>%s</Version>
-			<DeviceId>%s</DeviceId>
-			<MessageId>%s</MessageId>
-			<TimeStamp>%s</TimeStamp>
-			<ErrorCode>0</ErrorCode>
-			<ServiceStandbyMode>false</ServiceStandbyMode>
-			<Balance>
-				<Amount>2018</Amount>
-				<Currency>POINTS</Currency>
-			</Balance>
-			<Transactions>
-				<TransactionId>00000000</TransactionId>
-				<Date>%s</Date>
-				<Type>PURCHGAME</Type>
-			</Transactions>
-			<SyncTime>%s</SyncTime>
-			<ETickets>00000000</ETickets>
-			<Certs>00000000</Certs>
-			<Certs>00000000</Certs>
-			<TitleId>00000000</TitleId>
-		</PurchaseTitleResponse>
-	</soapenv:Body>
-</soapenv:Envelope>`, PT.Version, PT.DeviceID, PT.MessageID, timestamp, timestamp, timestamp)
-		fmt.Println("Delivered response!")
-
-	case "CheckRegistration":
-		fmt.Println("CR.")
-		CR := CR{}
-		if err = xml.Unmarshal(body, &CR); err != nil {
-			fmt.Println("...or not. Bad or incomplete request. (End processing.)")
-			fmt.Fprint(w, "not good enough for me. ;3")
-			fmt.Printf("Error: %s", err.Error())
-			return
-		}
-		fmt.Println(CR)
-		fmt.Println("The request is valid! Responding...")
-		fmt.Fprintf(w, `<?xml version="1.0" encoding="utf-8"?>
-<soapenv:Envelope xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/" 
-		xmlns:xsd="http://www.w3.org/2001/XMLSchema" 
-		xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-<soapenv:Body>
-<CheckRegistrationResponse xmlns="urn:ias.wsapi.broadon.com">
-	<Version>%s</Version>
-	<DeviceId>%s</DeviceId>
-	<MessageId>%s</MessageId>
-	<TimeStamp>%s</TimeStamp>
-	<ErrorCode>0</ErrorCode>
-	<ServiceStandbyMode>false</ServiceStandbyMode>
-	<OriginalSerialNumber>%s</OriginalSerialNumber>
-	<DeviceStatus>R</DeviceStatus>
-</CheckRegistrationResponse>
-</soapenv:Body>
-</soapenv:Envelope>`, CR.Version, CR.DeviceID, CR.DeviceID, timestamp, CR.SerialNo)
-		fmt.Println("[i] Delivered response!")
-
-	case "GetRegistrationInfo":
-		fmt.Println("GRI.")
-		GRI := GRI{}
-		if err = xml.Unmarshal(body, &GRI); err != nil {
-			fmt.Println("...or not. Bad or incomplete request. (End processing.)")
-			fmt.Fprint(w, "how dirty. ;3")
-			fmt.Printf("Error: %s", err.Error())
-			return
-		}
-		fmt.Println(GRI)
-		fmt.Println("The request is valid! Responding...")
-		fmt.Fprintf(w, `<?xml version="1.0" encoding="utf-8"?>
-<soapenv:Envelope   xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/" 
-		xmlns:xsd="http://www.w3.org/2001/XMLSchema" 
-		xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-	<soapenv:Body>
-		<GetRegistrationInfoResponse xmlns="urn:ias.wsapi.broadon.com">
-			<Version>%s</Version>
-			<DeviceId>%s</DeviceId>
-			<MessageId>%s</MessageId>
-			<TimeStamp>%s</TimeStamp>
-			<ErrorCode>0</ErrorCode>
-			<ServiceStandbyMode>false</ServiceStandbyMode>
-			<AccountId>%s</AccountId>
-			<DeviceToken>00000000</DeviceToken>
-			<DeviceTokenExpired>false</DeviceTokenExpired>
-			<Country>%s</Country>
-			<ExtAccountId></ExtAccountId>
-			<DeviceCode>0000000000000000</DeviceCode>
-			<DeviceStatus>R</DeviceStatus>
-			<Currency>POINTS</Currency>
-		</GetRegistrationInfoResponse>
-	</soapenv:Body>
-</soapenv:Envelope>`, GRI.Version, GRI.DeviceID, GRI.MessageID, timestamp, GRI.AccountID, GRI.Country)
-		fmt.Println("Delivered response!")
-
-	case "Register":
-		fmt.Println("REG.")
-		REG := REG{}
-		if err = xml.Unmarshal(body, &REG); err != nil {
-			fmt.Println("...or not. Bad or incomplete request. (End processing.)")
-			fmt.Fprint(w, "disgustingly invalid. ;3")
-			fmt.Printf("Error: %s", err.Error())
-			return
-		}
-		fmt.Println(REG)
-		fmt.Println("The request is valid! Responding...")
-		fmt.Fprintf(w, `<?xml version="1.0" encoding="utf-8"?>
-<soapenv:Envelope xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/" 
-		xmlns:xsd="http://www.w3.org/2001/XMLSchema" 
-		xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-	<soapenv:Body>
-		<RegisterResponse xmlns="urn:ias.wsapi.broadon.com">
-			<Version>%s</Version>
-			<DeviceId>%s</DeviceId>
-			<MessageId>%s</MessageId>
-			<TimeStamp>%s</TimeStamp>
-			<ErrorCode>0</ErrorCode>
-			<ServiceStandbyMode>false</ServiceStandbyMode>
-			<AccountId>%s</AccountId>
-			<DeviceToken>00000000</DeviceToken>
-			<Country>%s</Country>
-			<ExtAccountId></ExtAccountId>
-			<DeviceCode>00000000</DeviceCode>
-		</RegisterResponse>
-	</soapenv:Body>
-</soapenv:Envelope>`, REG.Version, REG.DeviceID, REG.MessageID, timestamp, REG.AccountID, REG.Country)
-		fmt.Println("Delivered response!")
-
-	case "Unregister":
-		fmt.Println("UNR.")
-		UNR := UNR{}
-		if err = xml.Unmarshal(body, &UNR); err != nil {
-			fmt.Println("...or not. Bad or incomplete request. (End processing.)")
-			fmt.Fprint(w, "how abnormal... ;3")
-			fmt.Printf("Error: %s", err.Error())
-			return
-		}
-		fmt.Println(UNR)
-		fmt.Println("The request is valid! Responding...")
-		fmt.Fprintf(w, `<?xml version="1.0" encoding="utf-8"?>
-<soapenv:Envelope xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/" xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-	<soapenv:Body>
-		<UnregisterResponse xmlns="urn:ias.wsapi.broadon.com">
-			<Version>%s</Version>
-			<DeviceId>%s</DeviceId>
-			<MessageId>%s</MessageId>
-			<TimeStamp>%s</TimeStamp>
-			<ErrorCode>0</ErrorCode>
-		<ServiceStandbyMode>false</ServiceStandbyMode>
-		</UnregisterResponse>
-	</soapenv:Body>
-</soapenv:Envelope>`, UNR.Version, UNR.DeviceID, UNR.MessageID, timestamp)
-		fmt.Println("Delivered response!")
-
-	default:
-		fmt.Fprintf(w, "WiiSOAP can't handle this. Try again later or actually use a Wii instead of a computer.")
-	}
-
-	// TODO: Add NUS and CAS SOAP to the case list.
-	fmt.Println("[!] End of Request." + "\n")
-}
-
-func namespaceForType(service string) string {
-	return "urn:" + service + ".wsapi.broadon.com"
-}
-
-// Expected contents are along the lines of "urn:ecs.wsapi.broadon.com/CheckDeviceStatus"
-func parseAction(original string, service string) string {
-	prefix := namespaceForType(service) + "/"
-	stripped := strings.Replace(original, prefix, "", 1)
-
-	if stripped == original {
-		// This doesn't appear valid.
-		return ""
-	} else {
-		return stripped
-	}
+	// From here on out, all special cool things should go into their respective handler function.
 }
